--- src/views/ImagePanel.java
@@ -1,30 +1,0 @@
-package views;
-import javax.swing.ImageIcon;
-import javax.swing.JLabel;
-import javax.swing.JPanel;
-import java.awt.*;
-
-class ImagePanel extends JPanel {
-
-    private Image img;
-  
-    public ImagePanel(String img) {
-      this(new ImageIcon(img).getImage());
-    }
-  
-    public ImagePanel(Image img) {
-      this.img = img;
-      Dimension size = new Dimension(img.getWidth(null), img.getHeight(null));
-      setPreferredSize(size);
-      setMinimumSize(size);
-      setMaximumSize(size);
-      setSize(size);
-      //setLayout(null);
-    }
-  
-    @Override
-    public void paintComponent(Graphics g) {
-      g.drawImage(img, 0, 0, null);
-    }
-  }
-
