+++ src/views/panel/MilitaryBuildingPanel.java
@@ -1,29 +1,16 @@
-package views;
+package views.panel;
 
 import java.awt.FontFormatException;
+import java.awt.GridLayout;
 import java.awt.event.ActionListener;
 import java.io.IOException;
-import javax.swing.Action;
-import javax.swing.JPanel;
+
+import javax.swing.JLabel;
 import javax.swing.JTextArea;
-import javax.swing.border.Border;
-import javax.swing.border.EmptyBorder;
 
 import engine.City;
-
-import java.awt.*;
-import java.awt.GridBagLayout;
-import javax.swing.JButton;
-import javax.swing.JFrame;
-import javax.swing.JLabel;
-import javax.swing.JPanel;
-import javax.swing.SwingUtilities;
-import javax.swing.border.EmptyBorder;
-
-import java.awt.Color;
-import java.awt.FontFormatException;
-import java.awt.event.*;
-import java.io.IOException;
+import views.button.CityButton;
+import views.button.StyledButton;
 
 public class MilitaryBuildingPanel extends BuildingPanel {
   private StyledButton recruit;
