+++ src/views/view/StyledLabel.java
@@ -1,4 +1,4 @@
-package views;
+package views.view;
 
 import javax.swing.JFrame;
 import javax.swing.JLabel;
