+++ src/views/view/WorldMapView.java
@@ -1,4 +1,4 @@
-package views;
+package views.view;
 
 import static javax.swing.ScrollPaneConstants.HORIZONTAL_SCROLLBAR_ALWAYS;
 import static javax.swing.ScrollPaneConstants.VERTICAL_SCROLLBAR_NEVER;
@@ -6,22 +6,19 @@ import static javax.swing.ScrollPaneConstants.VERTICAL_SCROLLBAR_NEVER;
 import java.awt.BorderLayout;
 import java.awt.Color;
 import java.awt.FontFormatException;
+import java.awt.event.ActionListener;
 import java.io.IOException;
-import java.awt.event.*;
 
-import javax.swing.Action;
 import javax.swing.BoxLayout;
 import javax.swing.ImageIcon;
-import javax.swing.JButton;
 import javax.swing.JFrame;
-import javax.swing.JLabel;
 import javax.swing.JPanel;
 import javax.swing.JScrollPane;
-import javax.swing.SwingConstants;
-import javax.swing.border.Border;
-import javax.swing.border.EmptyBorder;
 import javax.swing.plaf.DimensionUIResource;
-import javax.swing.text.AttributeSet.ColorAttribute;
+
+import views.button.CityButton;
+import views.panel.ImagePanel;
+import views.panel.PlayerPanel;
 
 public class WorldMapView extends JFrame {
   private static final String COLOR_BEIGE = "#C8AE81";
